# apps/docs/public/openapi/cyclist-profile.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cyclist Profile API",
    "version": "1.0.0",
    "description": "API for managing cyclist profiles",
    "contact": {
      "name": "Atlas Team",
      "url": "https://github.com/Ameciclo/atlas"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.ameciclo.org/cyclist-profile",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "Cyclist Profiles",
      "description": "Operations related to cyclist profiles"
    },
    {
      "name": "Cyclist Analytics",
      "description": "Analytics and insights for cyclist data"
    },
    {
      "name": "System",
      "description": "System operations"
    }
  ],
  "components": {
    "schemas": {},
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Health check response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "service": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "service"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles": {
      "get": {
        "tags": [
          "Cyclist Profiles"
        ],
        "responses": {
          "200": {
            "description": "The list of cyclist profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": -2147483648,
                        "maximum": 2147483647
                      },
                      "data": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "nullable": true
                          },
                          {
                            "type": "object",
                            "additionalProperties": {
                              "nullable": true
                            }
                          },
                          {
                            "type": "array",
                            "items": {
                              "nullable": true
                            }
                          },
                          {
                            "nullable": true
                          }
                        ]
                      },
                      "metadata": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          },
                          {
                            "nullable": true
                          },
                          {
                            "type": "object",
                            "additionalProperties": {
                              "nullable": true
                            }
                          },
                          {
                            "type": "array",
                            "items": {
                              "nullable": true
                            }
                          },
                          {
                            "nullable": true
                          }
                        ]
                      },
                      "coordinates": {
                        "nullable": true
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "updated_at": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "data",
                      "metadata",
                      "created_at",
                      "updated_at"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/{id}": {
      "get": {
        "tags": [
          "Cyclist Profiles"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "required": [
                "id"
              ],
              "example": 42
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested cyclist profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": -2147483648,
                      "maximum": 2147483647
                    },
                    "data": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "nullable": true
                        },
                        {
                          "type": "object",
                          "additionalProperties": {
                            "nullable": true
                          }
                        },
                        {
                          "type": "array",
                          "items": {
                            "nullable": true
                          }
                        },
                        {
                          "nullable": true
                        }
                      ]
                    },
                    "metadata": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "nullable": true
                        },
                        {
                          "type": "object",
                          "additionalProperties": {
                            "nullable": true
                          }
                        },
                        {
                          "type": "array",
                          "items": {
                            "nullable": true
                          }
                        },
                        {
                          "nullable": true
                        }
                      ]
                    },
                    "coordinates": {
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "data",
                    "metadata",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Cyclist profile not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "example": {
                    "message": "Not Found"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid id error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "path"
                            ]
                          }
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "issues",
                        "name"
                      ],
                      "example": {
                        "issues": [
                          {
                            "code": "invalid_type",
                            "expected": "number",
                            "received": "nan",
                            "path": [
                              "id"
                            ],
                            "message": "Expected number, received nan"
                          }
                        ],
                        "name": "ZodError"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/nearby": {
      "get": {
        "tags": [
          "Cyclist Profiles"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "lat",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "lon",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "default": 1000
            },
            "required": false,
            "name": "radius",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cyclist profiles within radius",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "path"
                            ]
                          }
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "issues",
                        "name"
                      ],
                      "example": {
                        "issues": [
                          {
                            "code": "invalid_type",
                            "expected": "number",
                            "received": "nan",
                            "path": [
                              "lat"
                            ],
                            "message": "Expected number, received nan"
                          },
                          {
                            "code": "invalid_type",
                            "expected": "number",
                            "received": "nan",
                            "path": [
                              "lon"
                            ],
                            "message": "Expected number, received nan"
                          }
                        ],
                        "name": "ZodError"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/nearby-summary": {
      "get": {
        "tags": [
          "Cyclist Profiles"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "lat",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "lon",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "default": 1000
            },
            "required": false,
            "name": "radius",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "nullable": true,
              "default": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cyclist profiles summary within radius",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number"
                    },
                    "profiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "gender": {
                            "type": "string",
                            "nullable": true
                          },
                          "age": {
                            "type": "number",
                            "nullable": true
                          },
                          "days_per_week": {
                            "type": "number",
                            "nullable": true
                          },
                          "bike_type": {
                            "type": "string",
                            "nullable": true
                          },
                          "neighborhood": {
                            "type": "string",
                            "nullable": true
                          },
                          "survey_year": {
                            "type": "string",
                            "nullable": true
                          },
                          "distance_meters": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "gender",
                          "age",
                          "days_per_week",
                          "bike_type",
                          "neighborhood",
                          "survey_year",
                          "distance_meters"
                        ]
                      }
                    }
                  },
                  "required": [
                    "total",
                    "profiles"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "path"
                            ]
                          }
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "issues",
                        "name"
                      ],
                      "example": {
                        "issues": [
                          {
                            "code": "invalid_type",
                            "expected": "number",
                            "received": "nan",
                            "path": [
                              "lat"
                            ],
                            "message": "Expected number, received nan"
                          },
                          {
                            "code": "invalid_type",
                            "expected": "number",
                            "received": "nan",
                            "path": [
                              "lon"
                            ],
                            "message": "Expected number, received nan"
                          }
                        ],
                        "name": "ZodError"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/summary": {
      "get": {
        "tags": [
          "Cyclist Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "year",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cyclist profile summary statistics",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          },
          "422": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string"
                              },
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "path"
                            ]
                          }
                        },
                        "name": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "issues",
                        "name"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/trends": {
      "get": {
        "tags": [
          "Cyclist Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "default": "2015,2018,2021,2024"
            },
            "required": false,
            "name": "years",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cyclist profile trends across years",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/gender-analysis": {
      "get": {
        "tags": [
          "Cyclist Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "year",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Gender-based cyclist analysis",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/safety-analysis": {
      "get": {
        "tags": [
          "Cyclist Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "year",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Safety analysis for cyclists",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/v1/cyclist-profiles/survey-locations": {
      "get": {
        "tags": [
          "Cyclist Analytics"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "nullable": true
            },
            "required": false,
            "name": "year",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "gender",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "race",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "income",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Survey locations with aggregated data",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    }
  }
}